+++ changelog.md
@@ -3,6 +3,13 @@
 ## UNRELEASED
 
 _Nothing yet..._
+**Breaking:**
+- bpk-component-tooltip:
+  - The `className` prop has been renamed to `portalClassName`.
+
+**Added:**
+- bpk-component-tooltip:
+  - Add `portalStyle` prop that enables styling the tooltip portal.
 
 ## 2017-08-03 - Deprecated Heading & Paragraph components